# doc/API_Documentation.openapi.yaml
openapi: 3.1.0
...
info:
  title: Core User Management API
  version: 1.0.0-draft
  summary: Reusable OpenAPI 3.1 contract for account lifecycle, authentication, user profile management, MFA, passkeys, admin user operations, and privacy preferences.
  description: |
    Canonical OpenAPI 3.1 contract for the reusable core user-management layer in this repository.

    This spec intentionally excludes the PriceTracker business-domain APIs such as catalog, shops, prices,
    purchases, comparisons, watchlists, alerts, and moderation of price submissions.

    The retained surface is limited to reusable platform concerns:

    - registration, login, logout, refresh, password reset, and MFA completion
    - current-user profile, avatar attachment, and contact management
    - passkeys, TOTP MFA, recovery codes, and security summary
    - account-owned private file handling needed by the core, such as avatars or export packages
    - admin user management and system settings
    - legal document discovery, privacy consent capture, and cookie preferences

    Contract conventions:

    - IDs use `format: uuid`
    - timestamps use `format: date-time` in UTC
    - default currency uses ISO 4217 alpha-3 codes
    - response bodies use `camelCase`
    - successful responses use `{ data, meta }`
    - error responses use `{ error }` and include `urgencyLevel` on a 1-9 scale
    - request DTOs reject unknown fields unless explicitly stated otherwise
servers:
...
    description: Production
...
paths:
...
  /admin/roles:
    get:
      tags:
        - AdminSecurity
      operationId: listRoles
      summary: List roles and permissions exposed by the core user system
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Role definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDefinitionListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/overview:
    get:
      tags:
        - AdminSettings
      operationId: getAdminOverview
      summary: Load admin overview counts for the core user system
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Core admin overview data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverviewEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/audit-logs:
    get:
      tags:
        - AdminSecurity
      operationId: listAuditLogs
      summary: List audit log entries
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/security/events:
    get:
      tags:
        - AdminSecurity
      operationId: listSecurityEvents
      summary: List security events across accounts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: Security event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSecurityEventListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users:
    get:
      tags:
        - AdminUsers
      operationId: listAdminUsers
      summary: List users for admin review and operations
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/RoleCodeQuery'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Admin user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserSummaryListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - AdminUsers
      operationId: createAdminManagedUser
      summary: Create a user or admin account from the admin console
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserSummaryEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/users/{accountId}:
    get:
      tags:
        - AdminUsers
      operationId: getAdminManagedUser
      summary: Get one user account for admin review
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: User summary for admin review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserSummaryEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - AdminUsers
      operationId: updateAdminManagedUser
      summary: Update one user account from the admin workspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
      responses:
        '200':
          description: Updated user summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserSummaryEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{accountId}/sessions:
    get:
      tags:
        - AdminSecurity
      operationId: listAdminUserSessions
      summary: List active sessions for one user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Session list for the selected user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{accountId}/sessions/revoke-all:
    post:
      tags:
        - AdminSecurity
      operationId: revokeAdminUserSessions
      summary: Revoke all sessions for one user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionBulkRevokeRequest'
      responses:
        '200':
          description: User sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{accountId}/security-events:
    get:
      tags:
        - AdminSecurity
      operationId: listAdminUserSecurityEvents
      summary: List security events for one user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Security events for the selected user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSecurityEventListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{accountId}/audit-logs:
    get:
      tags:
        - AdminSecurity
      operationId: listAdminUserAuditLogs
      summary: List audit log entries for one user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Audit log entries for the selected user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{accountId}/emails/{emailId}/verify:
    post:
      tags:
        - AdminSecurity
      operationId: adminVerifyUserEmail
      summary: Mark one user email as verified
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
        - $ref: '#/components/parameters/EmailIdPath'
      responses:
        '200':
          description: Updated verified email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAddressEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{accountId}/emails/{emailId}/unverify:
    post:
      tags:
        - AdminSecurity
      operationId: adminUnverifyUserEmail
      summary: Mark one user email as unverified
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
        - $ref: '#/components/parameters/EmailIdPath'
      responses:
        '200':
          description: Updated unverified email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAddressEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/bulk-actions:
    post:
      tags:
        - AdminUsers
      operationId: bulkAdminUserAction
      summary: Apply a bulk user action
      description: Apply a bulk action such as freeze, suspend, activate, restore, delete, or set-status to multiple accounts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserBulkActionRequest'
      responses:
        '200':
          description: Bulk action accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcknowledgementEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/settings:
    get:
      tags:
        - AdminSettings
      operationId: listSystemSettings
      summary: List persisted system settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System setting list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSystemSettingListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/settings/{settingKey}:
    patch:
      tags:
        - AdminSettings
      operationId: updateSystemSetting
      summary: Update one persisted system setting
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SettingKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSystemSettingUpdateRequest'
      responses:
        '200':
          description: Updated system setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSystemSettingEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    SearchQuery:
      name: query
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 120
      example: alex@example.com
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 256
      example: cur_eyJvZmZzZXQiOjIwfQ
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    StatusQuery:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum:
          - pending
          - active
          - email_unverified
          - password_reset_required
          - suspended
          - frozen
          - deleted
      example: active
    RoleCodeQuery:
      name: role
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 80
      example: admin
...
    EmailIdPath:
      name: emailId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
    AccountIdPath:
      name: accountId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: One or more fields are invalid.
                  urgencyLevel: 3
                  details:
                    field: email
                  requestId: req_01HZZZZZZZZZZ
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication is required.
                  urgencyLevel: 4
                  requestId: req_01HZZZZZZZZZZ
    Forbidden:
      description: The caller lacks permission for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: FORBIDDEN
                  message: You do not have permission to perform this action.
                  urgencyLevel: 5
                  requestId: req_01HZZZZZZZZZZ
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: NOT_FOUND
                  message: Resource not found.
                  urgencyLevel: 2
                  requestId: req_01HZZZZZZZZZZ
    Conflict:
      description: The request conflicts with existing state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: CONFLICT
                  message: The requested change conflicts with existing state.
                  urgencyLevel: 5
                  requestId: req_01HZZZZZZZZZZ
...
  schemas:
    Identifier:
      type: string
      format: uuid
      example: 11111111-1111-4111-8111-111111111111
...
    AcknowledgementEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Acknowledgement'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    EmailAddressEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/EmailAddress'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    AdminOverviewEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/AdminOverview'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    AdminUserSummaryEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/AdminUserSummary'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    AdminUserSummaryListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserSummary'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    AdminUserCreateRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - displayName
      properties:
        email:
          type: string
          format: email
          example: sam@example.com
        password:
          type: string
          minLength: 12
          maxLength: 128
          example: StrongPassword!234
        displayName:
          type: string
          minLength: 2
          maxLength: 80
          example: Sam Admin
        primaryPhone:
          type:
            - string
            - 'null'
          pattern: ^\+[1-9]\d{6,14}$
          example: '+447700900124'
        roleCodes:
          type: array
          items:
            type: string
          example:
            - user
        accountStatus:
          type: string
          enum:
            - pending
            - active
            - email_unverified
            - password_reset_required
            - suspended
            - frozen
          example: active
    AdminUserUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 80
          example: Sam Admin
        primaryEmail:
          type: string
          format: email
          example: sam@example.com
        primaryPhone:
          type:
            - string
            - 'null'
          pattern: ^\+[1-9]\d{6,14}$
          example: '+447700900124'
        roleCodes:
          type: array
          items:
            type: string
          example:
            - admin
        accountStatus:
          type: string
          enum:
            - pending
            - active
            - email_unverified
            - password_reset_required
            - suspended
            - frozen
            - deleted
          example: suspended
        requirePasswordChange:
          type: boolean
          example: true
        requireMfaEnrollment:
          type: boolean
          example: true
        disableLogin:
          type: boolean
          example: false
        defaultCurrency:
          $ref: '#/components/schemas/CurrencyCode'
        locale:
          type: string
          example: en-GB
        timezoneName:
          type: string
          example: Europe/London
        profileBio:
          type:
            - string
            - 'null'
          example: Updated by admin review.
        reason:
          type:
            - string
            - 'null'
          maxLength: 400
          example: Status changed after admin review.
...
    AdminSystemSettingListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminSystemSetting'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    SessionListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    SessionBulkRevokeRequest:
      type: object
      additionalProperties: false
      properties:
        scope:
          type: string
          enum:
            - others
            - all
          default: others
          example: others
        reason:
          type:
            - string
            - 'null'
          maxLength: 400
          example: Session rotation after suspicious activity.
...
    AdminSecurityEventListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminSecurityEvent'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    RoleDefinitionListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RoleDefinition'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    AuditLogListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        meta:
          $ref: '#/components/schemas/ResponseMeta'